+++ ha-ops/app/app_context.py
@@ -321,6 +321,9 @@ class AppContext:
     def core_reload_yaml(self):
         return supervisor.core_reload_yaml(self.call_supervisor)
 
+    def core_reload_lovelace(self):
+        return supervisor.core_reload_lovelace(self.call_supervisor)
+
     def do_core_check(self):
         return supervisor.do_core_check(self.call_supervisor)
 
@@ -718,6 +721,7 @@ class AppContext:
             clean_dir_names=policies.EXPORT_CLEAN_DIR_NAMES,
             clean_file_patterns=policies.EXPORT_CLEAN_FILE_PATTERNS,
             clean_paths=policies.EXPORT_CLEAN_PATHS,
+            core_reload_lovelace=self.core_reload_lovelace,
             core_restart=self.core_restart,
             core_reload_yaml=self.core_reload_yaml,
             core_start=self.core_start,
@@ -742,6 +746,7 @@ class AppContext:
             add_detail=self.add_detail,
             addon_action=self.addon_action,
             clear_tree=self.clear_tree,
+            core_reload_lovelace=self.core_reload_lovelace,
             core_reload_yaml=self.core_reload_yaml,
             core_restart=self.core_restart,
             core_start=self.core_start,